Move ConstraintsDSL class next to ConstraintParsers
